имитриуем работу loginApi слипами в mockLoginApi, в презентере используем метод логин
