--- app/src/main/java/com/example/c5/WebLoginApiImpl.kt
@@ -1,19 +1,0 @@
-package com.example.c5
-
-class WebLoginApiImpl:LoginApi {
-    override fun login(login: String, password: String): Boolean {
-        TODO("Not yet implemented")
-    }
-
-    override fun register(login: String, password: String, email: String): Boolean {
-        TODO("Not yet implemented")
-    }
-
-    override fun logout(): Boolean {
-        TODO("Not yet implemented")
-    }
-
-    override fun forgotPassword(login: String): Boolean {
-        TODO("Not yet implemented")
-    }
-}
