+++ app/src/main/java/com/example/c5/ui/login/LoginContract.kt
@@ -1,4 +1,4 @@
-package com.example.c5
+package com.example.c5.ui.login
 
 import androidx.annotation.MainThread
